moved signatureValidationMiddleware in the middleware folder
