+++ locksmith/__tests__/signatureValidationMiddleware.test.js
@@ -1,5 +1,5 @@
 var httpMocks = require('node-mocks-http')
-var signatureValidationMiddleware = require('../src/signatureValidationMiddleware')
+var signatureValidationMiddleware = require('../src/middlewares/signatureValidationMiddleware')
 
 let request, response
 let validSignature =
