+++ locksmith/src/app.js
@@ -1,7 +1,7 @@
 const bodyParser = require('body-parser')
 const cors = require('cors')
 const express = require('express')
-const signatureValidationMiddleware = require('./signatureValidationMiddleware')
+const signatureValidationMiddleware = require('./middlewares/signatureValidationMiddleware')
 
 const app = express()
 const router = express.Router()
